Recover UART1 & change chip index to 1 byte.

+++ firmware/be200_test.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 
 void prepare();
-void be200_base_test(uint16_t idx);
-void be200_mine(uint16_t idx);
-void be200_test(uint16_t idx);
+void be200_base_test(uint8_t idx);
+void be200_mine(uint8_t idx);
+void be200_test(uint8_t idx);
 void uart1_test();
 uint32_t be200_uart_nonce_test();
 void be200_uart_set_freq();
